Renamed raw.py to telemetry
